breaking up pipe types into separate files
